Consolidate misc helpers in new helper module
Remove py2exe build helper modules

Fix website URL in About dialog

Update installation instructions in README
