--- src/almostimportstdlib/almostimportstdlib.py
@@ -1,8 +1,0 @@
-# pylint: skip-file
-"""
-To understand this module,
-please refer to the containing package's documentation.
-"""
-
-if False:
-    pass
