--- src/misc/__init__.py
@@ -1,3 +1,0 @@
-"""
-A package of miscellaneous stuff for the PythonTurtle package.
-"""
